// src/types/order-item.ts
export interface OrderItem {
    id: string
    orderId: string
    productId: string
    productName: string
    categoryId: string
    categoryName: string
    gender: string
    size: string
    brand: string
    sku: string
    tags: string
    material: string
    quantity: number
    price: number
    totalAmount: number
}